crams-apps/crams_software/requirements.txt
pytz==2021.1
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
mysqlclient==2.0.3
rest_condition==1.0.3
gunicorn==20.0.4
urllib3==1.26.5
sqlparse==0.4.1
pandas==1.2.3
PyJWT==1.7.1